Fixed memory leak when there's a readdir clash
If a virtual BAM file is injected during the readdir loop before a
masking file, then the memory allocation for that virtual BAM will be
overwritten and lost. This commit fixes that by freeing any existing
key's memory -- which must be an injected file -- and deleting the key
from the hash table.